1. Removed static references to Mapper so that you can create your own instance of MappingServiceProvider and inject it, instead of having to reference the static class. 2. Added DNX451 project for compatibility with coreclr 3. Updated the cache key to be a long. While running coreclr (dnx451) on linux (mono) it was having issues with the existing cache key. With only taking 16 bits of each type's hash code, duplicate cache keys were being created. To fix this, I updated the cache key to 64 bits so that it can use all bits from source and destination types.

+++ ExpressMapper NET40/DestinationMappingService.cs
@@ -19,14 +19,14 @@ namespace ExpressMapper
 
         #region Privates
 
-        private readonly IDictionary<int, MulticastDelegate> _collectionMappers =
-            new Dictionary<int, MulticastDelegate>();
+        private readonly IDictionary<long, MulticastDelegate> _collectionMappers =
+            new Dictionary<long, MulticastDelegate>();
 
         #endregion
 
         #region Implementation of IMappingService
 
-        public override IDictionary<int, MulticastDelegate> CollectionMappers
+        public override IDictionary<long, MulticastDelegate> CollectionMappers
         {
             get { return _collectionMappers; }
         }
@@ -36,7 +36,7 @@ namespace ExpressMapper
             get { return true; }
         }
 
-        public override MulticastDelegate MapCollection(int cacheKey)
+        public override MulticastDelegate MapCollection(long cacheKey)
         {
             return CollectionMappers.ContainsKey(cacheKey) ? CollectionMappers[cacheKey] : null;
         }
